+++ src/Steps/FinalStep.tsx
@@ -1,25 +1,17 @@
 import { useContext } from 'react';
-import { Col, Form, Row } from 'react-bootstrap';
 import { globalContext } from '../stateManagement/GlobalStore';
 
 export const FinalStep = () => {
-  const { globalState, dispatch } = useContext(globalContext);
+  const { globalState } = useContext(globalContext);
   return (
-    <div>
-      <h1>Some more data entry, maybe based on prior data...</h1>
-
-      <Form.Group as={Row} className="mb-3" controlId="typeOfContract">
-        <Form.Label column sm={2}>Type of Contract</Form.Label>
-        <Col>
-          <Form.Select aria-label="Type of Contract select"
-            value={globalState.wizardOptions.type_of_contract}
-            onChange={e => dispatch({ type: 'MERGE_OPTION', payload: {'type_of_contract': e.target.value}})}>
-            <option>Firm Fixed Price</option>
-            <option>Cost Reimbursement</option>
-            <option>Time and Materials</option>
-          </Form.Select>
-        </Col>
-      </Form.Group>
+    <div className="m-3">
+      <h1>TMCR Template Steps Complete</h1>
+      {globalState.tmcrIndex === 0 && globalState.wizardOptions.length === 1 &&
+        <h3>You may add a second TMCR, or generate a document based on your inputs.</h3>}
+      {globalState.tmcrIndex === 0 && globalState.wizardOptions.length !== 1 &&
+        <h3>Activate the second TMCR in order to generate your document.</h3>}
+      {globalState.tmcrIndex === 1 && globalState.wizardOptions.length !== 1 &&
+        <h3>You may generate a document based on your inputs.</h3>}
     </div>
   );
 }