+++ src/stateManagement/reducer.ts
@@ -30,7 +30,6 @@ const Reducer = (state: GlobalStateInterface, action: ActionType): any => {
     case 'PURGE_STATE':
       return initialState;
     case 'ADD_TMCR':
-      // #TODO Future feature to run through wizard a second time for a 2nd TMCR 
       if (state.wizardOptions.length === 1) {
         let wizardOptions = [...state.wizardOptions];
         wizardOptions.push({...defaultOptions});